pyproject.toml
[build-system]
requires = [
    "setuptools>=60",
    "setuptools-scm>=8.0"
]
build-backend = "setuptools.build_meta"

[project]
name = "imagekit"
version = "0.2.0"
description = "A collection of image utility methods."
readme = "README.md"
requires-python = ">=3.9"
authors = [
    { name = "ajrlewis", email = "hello@ajrlewis.com" }
]
dependencies = [
    "numpy==1.25.2",
    "packaging==23.1",
    "Pillow==10.0.0",
    "pillow-heif==0.13.0",
    "pytesseract==0.3.10",
    "scipy==1.11.2",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"*" = ["*.*"]